Reorganized send tests
